refactor: combine processMedia and processImage - We will use processImage's logic for images, videos and audios
