+++ src/webhook/handlers/__tests__/processMedia.test.js
@@ -4,11 +4,8 @@ jest.mock('src/lib/detectDialogflowIntent');
 
 import MockDate from 'mockdate';
 import processMedia from '../processMedia';
-import {
-  oneIdenticalVideoArticle,
-  getVideoArticle,
-  notFound,
-} from '../__fixtures__/processMedia';
+import * as apiListArticlesResult from '../__fixtures__/processMedia';
+import * as apiGetArticleResult from '../__fixtures__/choosingArticle';
 import gql from 'src/lib/gql';
 import ga from 'src/lib/ga';
 
@@ -17,9 +14,9 @@ beforeEach(() => {
   gql.__reset();
 });
 
-it('one identical video article found and choose for user', async () => {
-  gql.__push(oneIdenticalVideoArticle);
-  gql.__push(getVideoArticle);
+it('one identical article found and choose for user', async () => {
+  gql.__push(apiListArticlesResult.oneIdenticalImageArticle);
+  gql.__push(apiGetArticleResult.oneImageArticle);
 
   const data = {
     sessionId: 1497994017447,
@@ -29,7 +26,7 @@ it('one identical video article found and choose for user', async () => {
     timestamp: 1497994016356,
     messageId: '6270464463537',
     message: {
-      type: 'video',
+      type: 'image',
       id: '6270464463537',
     },
   };
@@ -44,7 +41,7 @@ it('one identical video article found and choose for user', async () => {
         Object {
           "ea": "MessageType",
           "ec": "UserInput",
-          "el": "video",
+          "el": "image",
         },
       ],
       Array [
@@ -58,7 +55,7 @@ it('one identical video article found and choose for user', async () => {
         Object {
           "ea": "Search",
           "ec": "Article",
-          "el": "video-article-1",
+          "el": "image-article-1",
           "ni": true,
         },
       ],
@@ -66,7 +63,7 @@ it('one identical video article found and choose for user', async () => {
         Object {
           "ea": "Selected",
           "ec": "Article",
-          "el": "video-article-1",
+          "el": "image-article-1",
         },
       ],
       Array [
@@ -90,8 +87,58 @@ it('one identical video article found and choose for user', async () => {
   expect(ga.sendMock).toHaveBeenCalledTimes(2);
 });
 
-it('should handle video not found', async () => {
-  gql.__push(notFound);
+it('one article found (not identical)', async () => {
+  gql.__push(apiListArticlesResult.oneImageArticle);
+
+  const data = {
+    sessionId: 1497994017447,
+  };
+  const event = {
+    type: 'message',
+    timestamp: 1497994016356,
+    messageId: '6270464463537',
+    message: {
+      type: 'image',
+      id: '6270464463537',
+    },
+  };
+  const userId = 'Uc76d8ae9ccd1ada4f06c4e1515d46466';
+  MockDate.set('2020-01-01');
+  expect(await processMedia(data, event, userId)).toMatchSnapshot();
+  MockDate.reset();
+  expect(gql.__finished()).toBe(true);
+  expect(ga.eventMock.mock.calls).toMatchInlineSnapshot(`
+    Array [
+      Array [
+        Object {
+          "ea": "MessageType",
+          "ec": "UserInput",
+          "el": "image",
+        },
+      ],
+      Array [
+        Object {
+          "ea": "ArticleSearch",
+          "ec": "UserInput",
+          "el": "ArticleFound",
+        },
+      ],
+      Array [
+        Object {
+          "ea": "Search",
+          "ec": "Article",
+          "el": "image-article-1",
+          "ni": true,
+        },
+      ],
+    ]
+  `);
+  expect(ga.sendMock).toHaveBeenCalledTimes(1);
+});
+
+it('twelve articles found', async () => {
+  gql.__push(apiListArticlesResult.twelveImageArticles);
+
   const data = {
     sessionId: 1497994017447,
   };
@@ -100,7 +147,35 @@ it('should handle video not found', async () => {
     timestamp: 1497994016356,
     messageId: '6530038889933',
     message: {
-      type: 'video',
+      type: 'image',
+      id: '6530038889933',
+    },
+  };
+  const userId = 'Uc76d8ae9ccd1ada4f06c4e1515d46466';
+
+  MockDate.set('2020-01-01');
+  const result = await processMedia(data, event, userId);
+  MockDate.reset();
+  expect(result).toMatchSnapshot();
+  expect(gql.__finished()).toBe(true);
+  expect(result.replies.length).toBeLessThanOrEqual(5); // Reply message API limit
+  const carousel = result.replies.find(({ type }) => type === 'flex').contents;
+  expect(carousel.type).toBe('carousel');
+  expect(carousel.contents.length).toBeLessThanOrEqual(10); // Flex message carousel 10 bubble limit
+  expect(JSON.stringify(carousel).length).toBeLessThan(50 * 1000); // Flex message carousel 50K limit
+});
+
+it('should handle image not found', async () => {
+  gql.__push(apiListArticlesResult.notFound);
+  const data = {
+    sessionId: 1497994017447,
+  };
+  const event = {
+    type: 'message',
+    timestamp: 1497994016356,
+    messageId: '6530038889933',
+    message: {
+      type: 'image',
       id: '6530038889933',
     },
   };
@@ -115,7 +190,7 @@ it('should handle video not found', async () => {
         Object {
           "ea": "MessageType",
           "ec": "UserInput",
-          "el": "video",
+          "el": "image",
         },
       ],
       Array [
